Cleanup 11ty Config
