+++ package.json
@@ -22,8 +22,8 @@
 		"@tailwindcss/cli": "4.1.11",
 		"@tailwindcss/postcss": "4.1.11",
 		"@tailwindcss/typography": "0.5.16",
-		"@types/node": "24.0.10",
-		"luxon": "3.6.1",
+		"@types/node": "24",
+		"luxon": "3.7.1",
 		"markdown-it": "14.1.0",
 		"markdown-it-container": "4",
 		"postcss": "8.5.6",
